+++ src/core/currency/wallet/currency-wallet-pixie.ts
@@ -264,40 +264,87 @@ export const walletPixie: TamePixie<CurrencyWalletProps> = combinePixies({
       props.state.paused || props.walletState.paused ? undefined : props
   ),
 
-  syncNetwork: filterPixie(
+  syncNetworkUpdate: filterPixie(
+    (_input: CurrencyWalletInput) => {
+      return {
+        async update(props) {
+          const { engine } = props.walletOutput
+          if (engine?.syncNetwork == null) {
+            return
+          }
+          const { walletState } = props
+          const { currencyInfo, walletInfo, publicWalletInfo } = walletState
+          // Get the private keys if required by the engine:
+          const requiresPrivateKeys =
+            currencyInfo.unsafeSyncNetwork === true && publicWalletInfo != null
+          const privateKeys = requiresPrivateKeys ? walletInfo.keys : undefined
+          // Sync the network for each subscription:
+          for (const subscription of walletState.changeServiceSubscriptions) {
+            await engine.syncNetwork({
+              privateKeys,
+              subscribeParam: {
+                address: subscription.address,
+                checkpoint: subscription.checkpoint
+              }
+            })
+          }
+          // Update subscription status if managed by the change service:
+          props.dispatch({
+            type: 'CURRENCY_ENGINE_UPDATE_CHANGE_SERVICE_SUBSCRIPTIONS',
+            payload: {
+              subscriptions: walletState.changeServiceSubscriptions.map(
+                subscription => ({ ...subscription, status: 'listening' })
+              ),
+              walletId: props.walletId
+            }
+          })
+        },
+        destroy() {}
+      }
+    },
+    props =>
+      !props.state.paused &&
+      !props.walletState.paused &&
+      props.walletState.changeServiceSubscriptions.some(
+        subscription => subscription.status === 'syncing'
+      )
+        ? props
+        : undefined
+  ),
+  syncNetworkTask: filterPixie(
     (input: CurrencyWalletInput) => {
-      let syncNetworkTask: PeriodicTask
+      const syncNetworkTask: PeriodicTask = makePeriodicTask(
+        async (): Promise<void> => {
+          const { engine } = input.props.walletOutput
+          if (engine?.syncNetwork == null) {
+            syncNetworkTask.stop()
+            return
+          }
+          const { walletState } = input.props
+          const { currencyInfo, walletInfo, publicWalletInfo } = walletState
+          // Get the private keys if required by the engine:
+          const requiresPrivateKeys =
+            currencyInfo.unsafeSyncNetwork === true && publicWalletInfo != null
+          const privateKeys = requiresPrivateKeys ? walletInfo.keys : undefined
+          const delay = await engine.syncNetwork({
+            privateKeys
+          })
+          syncNetworkTask.setDelay(delay)
+        },
+        10000,
+        {
+          onError: error => {
+            input.props.log.error(error)
+          }
+        }
+      )
 
       return {
-        update() {
-          const { currencyInfo, walletInfo, publicWalletInfo } =
-            input.props.walletState
-          const { engine } = input.props.walletOutput
-          if (engine != null) {
+        update(props) {
+          const { engine } = props.walletOutput
+          if (engine?.syncNetwork != null && !syncNetworkTask?.started) {
             // Setup syncNetwork routine if defined by the currency engine:
-            if (engine.syncNetwork != null) {
-              // Get the private keys if required by the engine:
-              const requiresPrivateKeys =
-                currencyInfo.unsafeSyncNetwork === true &&
-                publicWalletInfo != null
-              const privateKeys = requiresPrivateKeys
-                ? walletInfo.keys
-                : undefined
-              const doNetworkSync = async (): Promise<void> => {
-                if (engine.syncNetwork != null) {
-                  const delay = await engine.syncNetwork({ privateKeys })
-                  syncNetworkTask.setDelay(delay)
-                } else {
-                  syncNetworkTask.stop()
-                }
-              }
-              syncNetworkTask = makePeriodicTask(doNetworkSync, 10000, {
-                onError: error => {
-                  input.props.log.error(error)
-                }
-              })
-              syncNetworkTask.start({ wait: false })
-            }
+            syncNetworkTask.start({ wait: false })
           }
         },
         destroy() {
@@ -309,7 +356,16 @@ export const walletPixie: TamePixie<CurrencyWalletProps> = combinePixies({
       }
     },
     props =>
-      props.state.paused || props.walletState.paused ? undefined : props
+      !props.state.paused &&
+      !props.walletState.paused &&
+      (props.walletState.changeServiceSubscriptions.length === 0 ||
+        props.walletState.changeServiceSubscriptions.some(
+          subscription =>
+            subscription.status === 'avoiding' ||
+            subscription.status === 'deferring'
+        ))
+        ? props
+        : undefined
   ),
 
   syncTimer: filterPixie(